Add files via upload
A script that searches the given string ( always with "minecraft:" at the beginning) in the current directory and all folders and subfolders just in .json files. If found, it prints the path. This is to find (in this case for minecraft json files) a item an every recipe, loot_table, advancement etc. if the default data folder of minecraft is given. This should help datapack creators.